# ollama-distributed/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OllamaMax Distributed API
  description: |
    OllamaMax Distributed is an enterprise-grade distributed AI platform that transforms 
    the single-node Ollama architecture into a horizontally scalable, fault-tolerant system.
    
    ## Authentication
    
    The API uses JWT-based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <JWT_TOKEN>
    ```
    
    ## Base URLs
    
    - **API Server**: `http://localhost:8080`
    - **Web Interface**: `http://localhost:8081`
    - **Metrics**: `http://localhost:9090`
    
    ## Rate Limiting
    
    Default rate limit: 1000 requests per minute per client.
    
  version: 1.0.0
  contact:
    name: OllamaMax Support
    url: https://github.com/khryptorgraphics/ollamamax
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.ollamamax.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      summary: User authentication
      description: Authenticate user and receive JWT token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "admin"
              password: "secure-password"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      summary: Refresh JWT token
      description: Get a new JWT token using a valid refresh token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/UserInfo'
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    UserInfo:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer
        details:
          type: string
        timestamp:
          type: string
          format: date-time
...